+++ judoscale-resque/test/metrics_collector_test.rb
@@ -19,23 +19,32 @@ module Judoscale
         subject.clear_queues
       }
 
-      it "collects latency for each queue" do
+      it "collects depth and latency for each queue" do
         queues = ["default", "high"]
         sizes = {"default" => 1, "high" => 2}
+        latencies = {"default" => 3.5, "high" => 4.999}
 
         metrics = ::Resque.stub(:queues, queues) {
           ::Resque.stub(:size, ->(queue_name) { sizes.fetch(queue_name) }) {
-            subject.collect
+            ::Resque.stub(:latency, ->(queue_name) { latencies.fetch(queue_name) }) {
+              subject.collect
+            }
           }
         }
 
-        _(metrics.size).must_equal 2
+        _(metrics.size).must_equal 4
         _(metrics[0].queue_name).must_equal "default"
         _(metrics[0].value).must_equal 1
         _(metrics[0].identifier).must_equal :qd
-        _(metrics[1].queue_name).must_equal "high"
-        _(metrics[1].value).must_equal 2
-        _(metrics[1].identifier).must_equal :qd
+        _(metrics[1].queue_name).must_equal "default"
+        _(metrics[1].value).must_equal 3500
+        _(metrics[1].identifier).must_equal :qt
+        _(metrics[2].queue_name).must_equal "high"
+        _(metrics[2].value).must_equal 2
+        _(metrics[2].identifier).must_equal :qd
+        _(metrics[3].queue_name).must_equal "high"
+        _(metrics[3].value).must_equal 4999
+        _(metrics[3].identifier).must_equal :qt
       end
 
       it "always collects for known queues" do
@@ -49,40 +58,48 @@ module Judoscale
 
         queues = ["default"]
         size = 0
+        latency = 0.0
 
         metrics = ::Resque.stub(:queues, queues) {
           ::Resque.stub(:size, size) {
-            subject.collect
+            ::Resque.stub(:latency, latency) {
+              subject.collect
+            }
           }
         }
 
-        _(metrics.size).must_equal 1
-        _(metrics[0].queue_name).must_equal "default"
+        _(metrics.size).must_equal 2
+        _(metrics.map(&:queue_name)).must_equal %w[default default]
 
         queues = []
 
         metrics = ::Resque.stub(:queues, queues) {
           ::Resque.stub(:size, size) {
-            subject.collect
+            ::Resque.stub(:latency, latency) {
+              subject.collect
+            }
           }
         }
 
-        _(metrics.size).must_equal 1
-        _(metrics[0].queue_name).must_equal "default"
+        _(metrics.size).must_equal 2
+        _(metrics.map(&:queue_name)).must_equal %w[default default]
       end
 
       it "logs debug information for each queue being collected" do
         use_config log_level: :debug do
           queues = ["default"]
           size = 2
+          latency = 3.0
 
           ::Resque.stub(:queues, queues) {
             ::Resque.stub(:size, size) {
-              subject.collect
+              ::Resque.stub(:latency, latency) {
+                subject.collect
+              }
             }
           }
 
-          _(log_string).must_match %r{resque-qd.default=2}
+          _(log_string).must_match %r{resque-qd.default=2 resque-qt.default=3000ms}
           _(log_string).wont_match %r{resque-busy}
         end
       end
@@ -91,6 +108,7 @@ module Judoscale
         use_adapter_config :resque, track_busy_jobs: true do
           queues = ["default", "high"]
           size = 2
+          latency = 1.0
 
           workers = [
             ResqueWorkerStub.new({"queue" => "default"}),
@@ -101,19 +119,21 @@ module Judoscale
 
           metrics = ::Resque.stub(:queues, queues) {
             ::Resque.stub(:size, size) {
-              ::Resque.stub(:working, workers) {
-                subject.collect
+              ::Resque.stub(:latency, latency) {
+                ::Resque.stub(:working, workers) {
+                  subject.collect
+                }
               }
             }
           }
 
-          _(metrics.size).must_equal 4
-          _(metrics[1].value).must_equal 2
-          _(metrics[1].queue_name).must_equal "default"
-          _(metrics[1].identifier).must_equal :busy
-          _(metrics[3].value).must_equal 1
-          _(metrics[3].queue_name).must_equal "high"
-          _(metrics[3].identifier).must_equal :busy
+          _(metrics.size).must_equal 6
+          _(metrics[2].value).must_equal 2
+          _(metrics[2].queue_name).must_equal "default"
+          _(metrics[2].identifier).must_equal :busy
+          _(metrics[5].value).must_equal 1
+          _(metrics[5].queue_name).must_equal "high"
+          _(metrics[5].identifier).must_equal :busy
         end
       end
 
@@ -122,17 +142,20 @@ module Judoscale
           use_adapter_config :resque, track_busy_jobs: true do
             queues = ["default"]
             size = 2
+            latency = 1.0
             workers = [ResqueWorkerStub.new({"queue" => "default"})]
 
             ::Resque.stub(:queues, queues) {
               ::Resque.stub(:size, size) {
-                ::Resque.stub(:working, workers) {
-                  subject.collect
+                ::Resque.stub(:latency, latency) {
+                  ::Resque.stub(:working, workers) {
+                    subject.collect
+                  }
                 }
               }
             }
 
-            _(log_string).must_match %r{resque-qd.default=2 resque-busy.default=1}
+            _(log_string).must_match %r{resque-qd.default=2 resque-qt.default=1000ms resque-busy.default=1}
           end
         end
       end
@@ -140,31 +163,39 @@ module Judoscale
       it "filters queues matching UUID format by default, to prevent reporting for dynamically generated queues" do
         queues = %W[low-#{SecureRandom.uuid} default #{SecureRandom.uuid}-high]
         size = 2
+        latency = 1.0
 
         metrics = ::Resque.stub(:queues, queues) {
           ::Resque.stub(:size, size) {
-            subject.collect
+            ::Resque.stub(:latency, latency) {
+              subject.collect
+            }
           }
         }
 
-        _(metrics.size).must_equal 1
-        _(metrics[0].queue_name).must_equal "default"
+        _(metrics.size).must_equal 2
+        _(metrics.map(&:queue_name)).must_equal %w[default default]
       end
 
       it "filters queues to collect metrics from based on the configured queue filter proc, overriding the default UUID filter" do
         use_adapter_config :resque, queue_filter: ->(queue_name) { queue_name.start_with? "low" } do
           queues = %W[low default high low-#{SecureRandom.uuid}]
           size = 2
+          latency = 1.0
 
           metrics = ::Resque.stub(:queues, queues) {
             ::Resque.stub(:size, size) {
-              subject.collect
+              ::Resque.stub(:latency, latency) {
+                subject.collect
+              }
             }
           }
 
-          _(metrics.size).must_equal 2
+          _(metrics.size).must_equal 4
           _(metrics[0].queue_name).must_equal "low"
-          _(metrics[1].queue_name).must_be :start_with?, "low-"
+          _(metrics[1].queue_name).must_equal "low"
+          _(metrics[2].queue_name).must_be :start_with?, "low-"
+          _(metrics[3].queue_name).must_be :start_with?, "low-"
         end
       end
 
@@ -172,14 +203,17 @@ module Judoscale
         use_adapter_config :resque, queues: %w[low ultra], queue_filter: ->(queue_name) { queue_name != "low" } do
           queues = %w[low default high]
           size = 2
+          latency = 1.0
 
           metrics = ::Resque.stub(:queues, queues) {
             ::Resque.stub(:size, size) {
-              subject.collect
+              ::Resque.stub(:latency, latency) {
+                subject.collect
+              }
             }
           }
 
-          _(metrics.map(&:queue_name)).must_equal %w[low ultra]
+          _(metrics.map(&:queue_name)).must_equal %w[low low ultra ultra]
         end
       end
 
@@ -187,14 +221,17 @@ module Judoscale
         use_adapter_config :resque, max_queues: 2 do
           queues = %w[low default high]
           size = 2
+          latency = 1.0
 
           metrics = ::Resque.stub(:queues, queues) {
             ::Resque.stub(:size, size) {
-              subject.collect
+              ::Resque.stub(:latency, latency) {
+                subject.collect
+              }
             }
           }
 
-          _(metrics.map(&:queue_name)).must_equal %w[low high]
+          _(metrics.map(&:queue_name)).must_equal %w[low low high high]
           _(log_string).must_match %r{Resque metrics reporting only 2 queues max, skipping the rest \(1\)}
         end
       end